[API] Minor migration changes.

+++ api/database/migrations/2022_07_11_225748_create_offerers_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
         Schema::create('offerers', function (Blueprint $table) {
             $table->string('id')->primary();
             $table->string('user_id');
-            $table->integer('rating')->default(0);
+            $table->float('rating')->default(0);
             $table->date('start_time')->useCurrent();
             $table->text('description');
             $table->bigInteger('jobs_completed')->default(0);
